requirements.txt
numpy
Pillow
opencv-python
git+https://github.com/opencv/opencv.git